Split compiler logic
